+++ docs_site/docs/.vitepress/theme/index.ts
@@ -1,4 +1,8 @@
 import DefaultTheme from "vitepress/theme";
+import AeLayout from "./AeLayout.vue";
 import "./custom.css";
 
-export default DefaultTheme;
+export default {
+  extends: DefaultTheme,
+  Layout: AeLayout,
+};